delivery_app/requirements.txt
Django
django-debug-toolbar
django-role-permissions
